factor out command code to seperate files
