+++ CLAUDE.md
@@ -32,6 +32,7 @@ Unit tests are in `src/event_library.rs` (`#[cfg(test)]` module) testing the nom
 
 ## Workflows
 - Rust code should be formatted as usual, but the c code in `src/bpf` should be formatted with `clang-format`.
+- Prefer the `foo.rs` + `foo/` style for modules with submodules, rather than `foo/mod.rs`.
 
 ## Gotchas
 - Never edit `src/bpf/sampler.skel.rs`. It is auto-generated by build.rs. If something looks wrong, do a `cargo clean && cargo build`.
